integration_tests/test-requirements.txt
git+https://github.com/rsichny/swagger-py.git  # the digium release does not support python3
https://github.com/wazo-pbx/ari-py/archive/0.4.0.zip
git+https://github.com/wazo-pbx/xivo-test-helpers.git
https://github.com/wazo-pbx/wazo-calld-client/archive/master.zip
https://github.com/wazo-pbx/xivo-lib-rest-client/archive/master.zip
docker-compose
kombu
nose
pyhamcrest
psycopg2-binary
requests
stevedore  # from wazo-calld-client
